+++ node/l1-bridge/src/event/daa_score_changed.rs
@@ -1,0 +1,6 @@
+/// Virtual DAA score changed.
+#[derive(Clone, Debug)]
+pub struct DaaScoreChanged {
+    /// New virtual DAA score.
+    pub daa_score: u64,
+}
